googlecloud/ml-invest-advisory/market_symbol_quotes/requirements.txt
shioaji
shioaji[speed]
google-cloud-secret-manager
functions-framework
PyYAML